refactor(reexecute): export NewMainnetCChainVM()
